# apps/web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TrekMind API",
    "description": "API do TrekMind para busca de destinos, lugares próximos e assistente de viagens. Autenticação via Better Auth (sessões/cookies para web) ou JWT Bearer (mobile).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento"
    }
  ],
  "tags": [
    { "name": "Auth", "description": "Autenticação (rotas de compatibilidade para mobile)" },
    { "name": "Chat", "description": "Assistente de viagens" },
    { "name": "Places", "description": "Busca de lugares e destinos" }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login (mobile)",
        "description": "Login por e-mail e senha. Retorna JWT para envio em Authorization: Bearer. Usado pelo app mobile.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Cadastro (mobile)",
        "description": "Cadastro de usuário por nome, e-mail e senha. Retorna JWT para envio em Authorization: Bearer.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": {
            "description": "E-mail já cadastrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimit" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": ["Chat"],
        "summary": "Perguntar ao assistente",
        "description": "Envia mensagem ao assistente de viagens. Requer autenticação (sessão ou Bearer).",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": { "type": "string" },
                  "latitude": { "type": "number", "minimum": -90, "maximum": 90 },
                  "longitude": { "type": "number", "minimum": -180, "maximum": 180 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta do assistente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "answer": { "type": "string" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/places/search": {
      "get": {
        "tags": ["Places"],
        "summary": "Buscar lugares",
        "description": "Busca lugares por termo (cidade, endereço, ponto de interesse). Requer autenticação.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Termo de busca" },
          { "name": "lat", "in": "query", "schema": { "type": "number" }, "description": "Latitude" },
          { "name": "lng", "in": "query", "schema": { "type": "number" }, "description": "Longitude" },
          { "name": "address", "in": "query", "schema": { "type": "string" } },
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "country", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de lugares",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Place" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/places/nearby": {
      "get": {
        "tags": ["Places"],
        "summary": "Lugares próximos",
        "description": "Busca lugares próximos a coordenadas. Requer autenticação.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "lat", "in": "query", "required": true, "schema": { "type": "number", "minimum": -90, "maximum": 90 } },
          { "name": "lng", "in": "query", "required": true, "schema": { "type": "number", "minimum": -180, "maximum": 180 } },
          { "name": "radius", "in": "query", "schema": { "type": "number", "default": 10 }, "description": "Raio em km" },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["restaurant", "museum", "beach", "trail", "hotel", "lodging", "bar", "nightlife", "park", "waterfall", "river", "shopping", "club", "water_park", "other"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de lugares",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Place" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT (mobile) ou sessão via cookie (web)"
      }
    },
    "schemas": {
      "Place": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "address": { "type": "string" },
          "source": { "type": "string", "enum": ["wikipedia", "google"] }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Não autorizado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "RateLimit": {
        "description": "Muitas requisições",
        "headers": { "Retry-After": { "schema": { "type": "integer" } } },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "InternalError": {
        "description": "Erro interno",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Conflict": {
        "description": "Conflito (ex: e-mail já cadastrado)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  }
}