# packages/application/src/dtos/place.dto.ts
export interface PlaceDTO {
  id: string;
  name: string;
  description?: string;
  category: string;
  latitude: number;
  longitude: number;
  address?: string;
  source?: string;
  url?: string;
  imageUrl?: string;
  wikipediaTitle?: string;
  createdAt: Date;
}